Setup initial unit testing framework
Create a sample unit test for `FilterButton`.
The general idea is that you load a component with some params, you render
it in a virtual screen and then you check the values of its attributes
and/or the presense of specific text.

Add `jsdom` dependency.

Move dev dependencies to `devDependencies` in `package.json`.

Modify `npm run test` script to run the tests.